requirements.txt
anyio==3.4.0
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
docopt==0.6.2
fastapi==0.70.1
ffmpeg-python==0.2.0
future==0.18.2
h11==0.12.0
idna==3.3
Js2Py==0.71
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pydantic==1.8.2
pyjsparser==2.7.1
pyparsing==3.0.6
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
starlette==0.16.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.16.0
